feat: support for ntfy authentication feat: support for more stripe event types
