+++ go.mod
@@ -1,5 +1,10 @@
 module github.com/vytskalt/stripe2ntfy
 
-go 1.21.9
+go 1.25
 
 require github.com/stripe/stripe-go/v79 v79.12.0
+
+require (
+	golang.org/x/net v0.23.0 // indirect
+	gopkg.in/yaml.v3 v3.0.0 // indirect
+)
